+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 # Change log
 
 ## v0.0.1
-- working ...
+- initial script and demo button